+++ CHANGELOG.md
@@ -4,6 +4,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Breaking changes:
+  - `GiftIconNext`: Remplaced by `ColorGiftIconNext`.
+  - `CrossCircleIconNext`: Remplaced by `ColorCrossCircleIconNext`.
+  - `HeartWithClickIconNext`: Remplaced by `ColorHeartWithClickIconNext`.
+- Feature: Add `CrossCircleIconNext` component.
+
 ## [7.3.0] - 2022-02-07
 
 Deprecations/Changes: